Move mocking for tests into separate module
